According to the Printify API documents, modify the data type of "width" and "height" form "int" to "double", modify the data type of "angle" form "double" to "int" and modify the method "CreateProductAsync" to produce request using "text/json".

+++ V1/Models/Shops/Products/Design.cs
@@ -27,10 +27,10 @@ namespace PrintifyApi.V1.Models.Shops.Products
         public string Type { get; set; }
 
         [JsonProperty("height")]
-        public int Height { get; set; }
+        public double Height { get; set; }
 
         [JsonProperty("width")]
-        public int Width { get; set; }
+        public double Width { get; set; }
 
     }
 }
